V1.5.1 - fixed warmth source radius check - build against bukkit 1.7.2 R0.1 DEV

+++ Arctica/src/com/github/CubieX/Arctica/ColdDamageEvent.java
@@ -8,16 +8,16 @@ public class ColdDamageEvent extends Event
 {
     private static final HandlerList handlers = new HandlerList();
     private static LivingEntity victim = null;
-    private static int damageToApply = 0;
+    private static double damageToApply = 0;
  
     //Constructor
-    public ColdDamageEvent(LivingEntity victim, int damageToApply)
+    public ColdDamageEvent(LivingEntity victim, double damageToApply)
     {
         ColdDamageEvent.victim = victim;
         ColdDamageEvent.damageToApply = damageToApply;
     }
     
-    public int getDamageToApply()
+    public double getDamageToApply()
     {
        return (damageToApply);
     }
